+++ shared/src/index.ts
@@ -2,3 +2,4 @@ export * from "./types.js";
 export * from "./formulas.js";
 export * from "./loot.js";
 export * from "./monsters.js";
+export * from "./achievements.js";